fix: fixed authentication and permissions

+++ teddy-backend/src/main/java/br/com/teddy/store/controller/CartController.java
@@ -1,7 +1,9 @@
 package br.com.teddy.store.controller;
 
 import br.com.teddy.store.domain.Item;
+import br.com.teddy.store.repostiory.ICustomerRepository;
 import br.com.teddy.store.service.ICartService;
+import br.com.teddy.store.service.ICustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -14,8 +16,14 @@ public class CartController {
     @Autowired
     ICartService cartService;
 
+    @Autowired
+    ICustomerService customerService;
+
     @PostMapping("")
     public ResponseEntity addCartItem(@PathVariable Long id, @RequestBody Item item) {
+        if(!customerService.isCurrentUserLoggedIn(id)) {
+            return ResponseEntity.status(403).body("");
+        }
         try {
             cartService.addCartItem(id, item);
         } catch (Exception e) {
@@ -27,6 +35,9 @@ public class CartController {
 
     @DeleteMapping("/{itemID}")
     public ResponseEntity removeCartItem(@PathVariable Long id, @PathVariable Long itemID) {
+        if(!customerService.isCurrentUserLoggedIn(id)) {
+            return ResponseEntity.status(403).body("");
+        }
         cartService.removeCartItem(id, itemID);
 
         return ResponseEntity.ok("ok");
@@ -34,6 +45,9 @@ public class CartController {
 
     @PatchMapping("")
     public ResponseEntity updateCartItemAmount(@PathVariable Long id, @RequestBody Item item) {
+        if(!customerService.isCurrentUserLoggedIn(id)) {
+            return ResponseEntity.status(403).body("");
+        }
         cartService.updateCartItemAmount(id, item);
         return ResponseEntity.ok("ok");
     }
